choose BigInteger Count for now

+++ src/main/kotlin/com/icosahedron/Count.kt
@@ -2,19 +2,16 @@ package com.icosahedron
 
 import java.math.BigInteger
 
-inline class Count(private val value: Long) {
-    constructor(value: BigInteger) : this(value.toLong())
-    constructor(value: Int) : this(value.toLong())
+inline class Count(private val value: BigInteger) {
+    constructor(value: Long) : this(value.toBigInteger())
+    constructor(value: Int) : this(value.toBigInteger())
     operator fun plus(other: Count) = Count(value + other.value)
     override fun toString() = value.toString()
 }
 
-//import java.math.BigInteger
-//
-//inline class BigIntegerCount(private val value: BigInteger) {
-//    constructor(value: Long) : this(value.toBigInteger())
-//    constructor(value: Int) : this(value.toBigInteger())
-//    operator fun plus(other: BigIntegerCount) = BigIntegerCount(value + other.value)
+//inline class Count(private val value: Long) {
+//    constructor(value: BigInteger) : this(value.toLong())
+//    constructor(value: Int) : this(value.toLong())
+//    operator fun plus(other: Count) = Count(value + other.value)
 //    override fun toString() = value.toString()
 //}
-
